rt: split up `runtime::context` into multiple files
