+++ tokio/src/runtime/context/current.rs
@@ -2,7 +2,6 @@ use super::{Context, CONTEXT};
 
 use crate::runtime::{scheduler, TryCurrentError};
 use crate::util::markers::SyncNotSend;
-use crate::util::rand::{FastRand, RngSeed};
 
 use std::marker::PhantomData;
 
@@ -10,15 +9,23 @@ use std::marker::PhantomData;
 #[must_use]
 pub(crate) struct SetCurrentGuard {
     old_handle: Option<scheduler::Handle>,
-    old_seed: RngSeed,
     _p: PhantomData<SyncNotSend>,
 }
 
 /// Sets this [`Handle`] as the current active [`Handle`].
 ///
 /// [`Handle`]: crate::runtime::scheduler::Handle
-pub(crate) fn try_set_current_guard(handle: &scheduler::Handle) -> Option<SetCurrentGuard> {
-    CONTEXT.try_with(|ctx| ctx.set_current_guard(handle)).ok()
+pub(crate) fn try_set_current(handle: &scheduler::Handle) -> Option<SetCurrentGuard> {
+    CONTEXT
+        .try_with(|ctx| {
+            let old_handle = ctx.handle.borrow_mut().replace(handle.clone());
+
+            SetCurrentGuard {
+                old_handle,
+                _p: PhantomData,
+            }
+        })
+        .ok()
 }
 
 pub(crate) fn with_current<F, R>(f: F) -> Result<R, TryCurrentError>
@@ -33,17 +40,11 @@ where
 }
 
 impl Context {
-    pub(super) fn set_current_guard(&self, handle: &scheduler::Handle) -> SetCurrentGuard {
-        let rng_seed = handle.seed_generator().next_seed();
-
+    pub(super) fn set_current(&self, handle: &scheduler::Handle) -> SetCurrentGuard {
         let old_handle = self.handle.borrow_mut().replace(handle.clone());
-        let mut rng = self.rng.get().unwrap_or_else(FastRand::new);
-        let old_seed = rng.replace_seed(rng_seed);
-        self.rng.set(Some(rng));
 
         SetCurrentGuard {
             old_handle,
-            old_seed,
             _p: PhantomData,
         }
     }
@@ -53,10 +54,6 @@ impl Drop for SetCurrentGuard {
     fn drop(&mut self) {
         CONTEXT.with(|ctx| {
             *ctx.handle.borrow_mut() = self.old_handle.take();
-
-            let mut rng = ctx.rng.get().unwrap_or_else(FastRand::new);
-            rng.replace_seed(self.old_seed.clone());
-            ctx.rng.set(Some(rng));
         });
     }
 }
