+++ tokio/src/util/markers.rs
@@ -2,3 +2,7 @@
 pub(crate) struct SyncNotSend(*mut ());
 
 unsafe impl Sync for SyncNotSend {}
+
+cfg_rt! {
+    pub(crate) struct NotSendOrSync(*mut ());
+}
